Improve HTTP mock / test utilities
